make run archiver a service
